Extract IfElse into separate struct
